seachBarComponent now uses node Search #526
